# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/transaction/index/{portfolio}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get portfolio transactions",
                "description": "Retrieve a list of transactions for a given portfolio",
                "operationId": "7f469a60243280d528b021f6e7413b6c",
                "parameters": [
                    {
                        "name": "portfolio",
                        "in": "path",
                        "description": "ID of the portfolio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/transaction/show/{transaction}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction details",
                "description": "Retrieve details of a specific transaction",
                "operationId": "2e5d83253b2e7df735bbfc7ed90d2bd1",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/transaction/store": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create a new transaction",
                "description": "Add a new transaction to the portfolio",
                "operationId": "d0460acf126fa17fb0017c5e5e54a11b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/transaction/update/{transaction}": {
            "patch": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update a transaction",
                "description": "Modify an existing transaction",
                "operationId": "4f85576226b25e27c7df99192073b06b",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "ID of the transaction to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/transaction/delete/{transaction}": {
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete a transaction",
                "description": "Remove a specific transaction from the portfolio",
                "operationId": "f9fbfae8470a87a24c4b5276a5021596",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "ID of the transaction to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/stats/transaction/get/{transaction}": {
            "get": {
                "tags": [
                    "Transaction Stats"
                ],
                "summary": "Get statistics for a specific transaction",
                "description": "Retrieve profit and loss statistics for a single transaction",
                "operationId": "5a2f5f9b5c3214e4ec4d9c096719f67b",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "profitValuePercent": {
                                            "description": "Percentage of profit/loss",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "profitValuePrice": {
                                            "description": "Absolute profit/loss value in USD",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "profitSide": {
                                            "description": "Price change direction (+/-)",
                                            "type": "string",
                                            "enum": [
                                                "+",
                                                "-"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "403": {
                        "description": "Insufficient rights to access the transaction"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "TransactionRequest": {
                "properties": {
                    "coin_name": {
                        "description": "Name of the cryptocurrency involved in the transaction",
                        "type": "string",
                        "maxLength": 30
                    },
                    "description": {
                        "description": "A description of the transaction",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "amount": {
                        "description": "Amount of the cryptocurrency involved in the transaction (required if total_value_in_usd is not provided)",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "total_value_in_usd": {
                        "description": "Total value of the transaction in USD (required if amount is not provided)",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "is_buying": {
                        "description": "Indicates whether the transaction is a buy (true) or sell (false)",
                        "type": "boolean"
                    },
                    "portfolio_id": {
                        "description": "ID of the portfolio associated with the transaction",
                        "type": "integer"
                    },
                    "transaction_date": {
                        "description": "Date and time when the transaction occurred (format: dd-mm-yyyy)",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "properties": {
                    "id": {
                        "description": "Id of transaction",
                        "type": "number"
                    },
                    "coin_name": {
                        "description": "Name of the cryptocurrency involved in the transaction",
                        "type": "string"
                    },
                    "description": {
                        "description": "A description of the transaction",
                        "type": "string"
                    },
                    "amount": {
                        "description": "Amount of the cryptocurrency involved in the transaction",
                        "type": "number",
                        "format": "float"
                    },
                    "price_at_buy_moment": {
                        "description": "Price of the cryptocurrency at the moment of the transaction",
                        "type": "number",
                        "format": "float"
                    },
                    "total_value_in_usd": {
                        "description": "Total value of the transaction in USD",
                        "type": "number",
                        "format": "float"
                    },
                    "is_buying": {
                        "description": "Indicates whether the transaction is a buy (true) or sell (false)",
                        "type": "boolean"
                    },
                    "portfolio_id": {
                        "description": "ID of the portfolio associated with the transaction",
                        "type": "integer"
                    },
                    "transaction_date": {
                        "description": "Date and time when the transaction occurred",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "SanctumAuth": {
                "type": "apiKey",
                "description": "Введите токен в формате 'Bearer {your-token}'",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}